Expose Paging class (#36)
* Expose Paging class

* Move to correct file
