+++ src/main/java/com/sigopt/model/Pagination.java
@@ -28,22 +28,3 @@ public class Pagination<T extends APIObject> extends StructObject {
     }
 }
 
-class Paging extends StructObject {
-    public Paging() {
-        super();
-    }
-
-    public Paging(String before, String after) {
-        super();
-        this.set("before", before);
-        this.set("after", after);
-    }
-
-    public String getBefore() {
-        return (String) this.get("before");
-    }
-
-    public String getAfter() {
-        return (String) this.get("after");
-    }
-}
